Refactor table_method to an enum Rename formulation to types
